Add Document Display and Edit components into their own files with corresponding test file change
